+++ __manifest__.py
@@ -12,7 +12,8 @@
     'data': [
         'data/account_report.xml',
         'report/report_account_standard_report.xml',
-        'view/account_view.xml',
+        'views/account_view.xml',
+        'views/account_standard.xml',
         'wizard/account_standard_report_view.xml',
     ],
     'demo': [],
